+++ wp-content/themes/appiappi-theme/functions.php
@@ -13,3 +13,6 @@ require get_template_directory() . '/inc/setup.php';
 require get_template_directory() . '/inc/enqueue.php';
 require get_template_directory() . '/inc/customizer.php';
 require get_template_directory() . '/inc/template-tags.php';
+require get_template_directory() . '/inc/admin/settings-page.php';
+require get_template_directory() . '/inc/seo.php';
+require get_template_directory() . '/inc/security.php';